backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
psutil==5.9.6
aiofiles==23.1.0
python-pam==2.0.2
python-jose[cryptography]==3.5.0
